add Favourites cog
Move the 'favourites' group of commands to its own cog, allowing the user to disable them, since they aren't that useful if
they are already using the Playlist cog.
