+++ .env.example
@@ -28,6 +28,7 @@ LOG_COUNT="7"
 # Cog selection - set an empty string to disable a cog, write something to enable it.
 # Disabling a cog means the variables for that cog aren't required.
 # Don't forget to add/set them when enabling the cog.
+ENABLE_JSONFAV="ON"
 ENABLE_MEME="ON"
 ENABLE_MISC="ON"
 ENABLE_MUSIC="ON"
