+++ cogs/music.py
@@ -34,9 +34,7 @@
 # DEALINGS IN THE SOFTWARE.
 
 
-import aiofiles
 import asyncio
-import json
 import logging
 import os
 
@@ -54,7 +52,7 @@ from cogs.ext.queue import SongQueue, EmptyQueueError
 from cogs.ext.song import Song
 
 from dotenv import load_dotenv
-from typing import Tuple, Union, List, Dict
+from typing import Tuple, Union
 
 
 # Colours and string for some coloured output
@@ -69,13 +67,8 @@ load_dotenv()
 MAX_DURATION = int(os.getenv('MAX_DURATION'))
 # Directory in which music is downloaded
 SAVE_DIR = os.getenv('MUSIC_DIR')
-# Name of the file in which the list of favourites is kept.
-# By default it's 'favourite_songs.json', the directory is rsc/
-FAV_LIST_FILE = f"rsc/{os.getenv('MUSIC_FAV_LIST', 'favourite_songs.json')}"
 
 BOT_PREFIX = os.getenv('BOT_PREFIX')
-
-FAV_LIST = None
 
 
 # Used to suppress useless errors apparently
@@ -888,330 +881,6 @@ class Music(commands.Cog):
 			await vc.move_to(user_channel)
 			self.info[gid]['channel'] = user_channel
 			await ctx.send(f"Moved the bot to {user_channel.name}")
-
-	@commands.group(
-		name="favourites",
-		aliases=["fav", "favorites"],
-		help="Base command for managing your list of favourite songs"
-	)
-	async def favourites(self, ctx: commands.Context):
-		"""
-		Base command for managing favourite songs.
-
-		Loads the list of favourites from FAV_LIST_FILE during the first call.
-		"""
-		global FAV_LIST
-
-		logger = self.logger
-
-		# First verify a subcommand has been used.
-		if ctx.invoked_subcommand is None:
-			await ctx.send("You have to use a subcommand:")
-			await ctx.send_help(self.favourites)
-			return
-
-		# Then load the lists if needed.
-		if FAV_LIST is None:
-
-			try:
-				async with aiofiles.open(FAV_LIST_FILE, 'r') as file:
-					content = await file.read()
-
-				if content:
-					FAV_LIST = json.loads(content)
-				else:
-					FAV_LIST = dict()
-
-			except IOError as ioe:
-				logger.warning(
-					f"Could not open \"{FAV_LIST_FILE}\", maybe the file doesn't exist: ignoring."
-				)
-				logger.debug(f"IOError:\n{ioe}")
-				FAV_LIST = dict()
-
-			except Exception as e:
-				logger.error(f"Could not load \"{FAV_LIST_FILE}\".")
-				logger.debug(f"Unexpected exception:\n{e}")
-				await ctx.send("A problem occurred while loading your list, please try again.")
-				return
-
-	@favourites.command(
-		name="list",
-		help="Displays your list of favourites songs: if an index is specified, shows that song's info"  # noqa: E501
-	)
-	async def fav_list(self, ctx: commands.Context, index: int = 0):
-		"""
-		Subcommand to display the list of an user's favourite songs.
-
-		Parameters:
-			index: If less than or equal to zero, it displays the whole list.
-			If greater than zero, it displays the song with that index in the list with more info.
-		"""
-		member: discord.Member = ctx.message.author
-		member_id: str = str(member.id)
-
-		songs = FAV_LIST.get(member_id, None)
-
-		if songs is not None:
-
-			if len(songs) == 0:
-
-				name = member.display_name
-				if name[-1] == 's':
-					title = f"{name}' list"
-				else:
-					title = f"{name}'s list"
-
-				em = discord.Embed(title=title, description="Your list is empty.")
-
-				await ctx.send(embed=em)
-				return
-
-			if index <= 0:
-
-				cpt = 1
-				message = ""
-				for song in songs:
-					message += f"{cpt}. {song.get('title')}\n"
-					cpt += 1
-
-				name = member.display_name
-				if name[-1] == 's':
-					title = f"{name}' list"
-				else:
-					title = f"{name}'s list"
-
-				em = discord.Embed(
-					title=title,
-					description=message,
-					colour=member.colour
-				)
-
-				await ctx.send(embed=em)
-
-			else:
-
-				if index > len(songs):
-
-					await ctx.send(f"There's no song with that index! Your list has {len(songs)} songs.")
-
-				else:
-
-					song = songs[index - 1]
-
-					name = member.display_name
-					if name[-1] == 's':
-						title = f"From {name}' list"
-					else:
-						title = f"From {name}'s list"
-
-					em = discord.Embed(title=title)
-					em.add_field(name="Title", value=song.get('title'), inline=False)
-					em.add_field(name="URL", value=song.get('url'))
-					em.set_thumbnail(url=song.get('thumbnail'))
-
-					await ctx.send(embed=em)
-
-		else:
-			await ctx.send(f"You haven't saved any songs yet, use `{BOT_PREFIX}favourites add <song URL>` or `{BOT_PREFIX}favourites now` to add one.")  # noqa: E501
-
-	@favourites.command(
-		name="add",
-		help="Saves a song to your list from its URL"
-	)
-	async def fav_add(self, ctx: commands.Context, url: str = None):
-		"""
-		Saves a song to the user's list: uses from_url to get the title and the thumbnail's URL.
-
-		Parameters:
-			url: The URL of the song to save.
-		"""
-		global FAV_LIST
-
-		logger = self.logger
-
-		if url is None:
-			await ctx.send("You have to provide a URL.")
-			return
-
-		try:
-			# Suppress the embed to avoid clutter.
-			msg: discord.Message = ctx.message
-			await msg.edit(suppress=True)
-
-			song = await YTDLSource.from_url(url, download=False)
-			info = dict()
-			info['title'] = song.title
-			info['url'] = song.url
-			info['thumbnail'] = song.thumbnail
-
-			member: discord.Member = ctx.message.author
-			member_id: str = str(member.id)
-
-			songs = FAV_LIST.get(member_id, None)
-
-			if songs is not None:
-				songs.append(info)
-			else:
-				songs = [info]
-				FAV_LIST[member_id] = songs
-
-			dump = json.dumps(FAV_LIST)
-			async with aiofiles.open(FAV_LIST_FILE, 'w') as file:
-				await file.write(dump)
-
-			em = discord.Embed(description=f"Added \"{info.get('title')}\" to your list.")
-
-			await ctx.send(embed=em)
-
-		except MaxDurationError:
-			await ctx.send("The song is too long to be played, so it can't be saved.")
-
-		except Exception as e:
-			logger.error("Could not add a song to list of favourites.")
-			logger.debug(f"Exception:\n{e}")
-			await ctx.send("An error occurred while saving the song, please try again.")
-
-	@favourites.command(
-		name="remove",
-		help="Removes a song from your list by its index, 0 means no song is removed"
-	)
-	async def fav_remove(self, ctx: commands.Context, index: int = 0):
-		"""
-		Remove a song from the user's list by its index.
-
-		Parameters:
-			index: The index of the song to remove, 0 by default to avoid removing anything.
-		"""
-		global FAV_LIST
-
-		logger = self.logger
-
-		if index <= 0:
-			await ctx.send(f"You have to provide a valid index. Use `{BOT_PREFIX}favourites list` to check your list.")  # noqa: E501
-			return
-
-		member_id: str = str(ctx.message.author.id)
-		songs: List[Dict[str, str]] = FAV_LIST.get(member_id, None)
-
-		if songs is None:
-			await ctx.send("You haven't saved any songs yet.")
-			return
-
-		if index > len(songs):
-			await ctx.send(
-				f"You have to select a valid index: there's only {len(songs)} songs in your list."
-			)
-			return
-
-		song = songs.pop(index - 1)
-
-		try:
-			dump = json.dumps(FAV_LIST)
-			async with aiofiles.open(FAV_LIST_FILE, 'w') as file:
-				await file.write(dump)
-
-			message = f"Removed \"{song.get('title')}\" from your list."
-			em = discord.Embed(description=message)
-			await ctx.send(embed=em)
-
-		except Exception as e:
-			logger.error("Couldn't save favourites list after removing a song.")
-			logger.debug(f"Exception:\n{e}")
-
-	@favourites.command(
-		name="now",
-		help="Saves the currently playing song to your list"
-	)
-	@commands.guild_only()
-	async def fav_now(self, ctx: commands.Context):
-		"""
-		Saves the currently playing song from self.info.source.
-		"""
-		global FAV_LIST
-
-		logger = self.logger
-
-		gid = ctx.message.guild.id
-
-		if gid not in self.info:
-			await ctx.send("The bot is not connected to a voice channel.")
-			return
-
-		info = self.info.get(gid)
-		source: YTDLSource = info.get('source')
-
-		if source is not None:
-
-			member_id: str = str(ctx.message.author.id)
-
-			songs = FAV_LIST.get(member_id, None)
-
-			song = dict()
-			song['title'] = source.title
-			song['url'] = source.url
-			song['thumbnail'] = source.thumbnail
-
-			if songs is not None:
-				songs.append(song)
-			else:
-				songs = [song]
-				FAV_LIST[member_id] = songs
-
-			try:
-				dump = json.dumps(FAV_LIST)
-				async with aiofiles.open(FAV_LIST_FILE, 'w') as file:
-					await file.write(dump)
-
-				em = discord.Embed(description=f"Added \"{song.get('title')}\"to your list.")
-
-				await ctx.send(embed=em)
-
-			except Exception as e:
-				logger.error("Could not save a song to FAV_LIST_FILE.")
-				logger.debug(f"Exception:\n{e}")
-				await ctx.send("An error occurred while saving the song, please try again.")
-
-		else:
-
-			await ctx.send(
-				f"The bot is not playing something at the moment, try `{BOT_PREFIX}play <song>`."
-			)
-
-	@favourites.command(
-		name="play",
-		help="Plays a song from your list by its index"
-	)
-	@commands.guild_only()
-	async def fav_play(self, ctx: commands.Context, index: int = 0):
-		"""
-		Play a song from the user's list by passing its URL to self.play.
-
-		Parameters:
-			index: The index of the song to play, 0 by default.
-		"""
-
-		if index <= 0:
-			await ctx.send("You have to provide a valid index.")
-			return
-
-		member_id: str = str(ctx.message.author.id)
-		songs = FAV_LIST.get(member_id, None)
-
-		if songs is None:
-			await ctx.send(f"You haven't saved a song yet, you can use `{BOT_PREFIX}favourites add <URL>` or `{BOT_PREFIX}favourites now`.")  # noqa: E501
-			return
-
-		if index > len(songs):
-			await ctx.send(
-				f"You have to provide a valid index, your list only has {len(songs)} songs."
-			)
-			return
-
-		song = songs[index - 1]
-
-		url = song.get('url')
-		await self.play(ctx, url)
 
 	@commands.Cog.listener('on_voice_state_update')
 	async def on_empty_channel(
